doro.sps: implemented `symlink-bundle' command
This command is useful in development, where you want to edit the
files in-place, and not continuosly re-install.

`symlink-bundle' does not unconditionally symlink each file
separately, but produces a shallower tree by also symlinking
directories if the result is semantically identical.

The link-each-file behaviour can be enabled by the `--deep' option.