Implement destination hooks
There is now a destination-specific list of hooks run after a package
has been unpacked and its installation hook run.

This facility is used by the FHS destination to create the shell
wrappers for programs, instead of doing that directly upon
installation of these programs.

A nice effect of this change is that each entry in an inventory of a
package maps to exactly one pathname for the destination, instead of
potentially multiple ones.